pyproject.toml
[tool.poetry]
name = "codereviewai"
version = "0.1.0"
description = "description"
authors = ["mdubyna <mykola22dubyna@gmail.com>"]
#readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.6"}
mypy = "^1.14.0"
black = "^24.10.0"
uvicorn = "^0.34.0"
pydantic = "^2.10.4"
redis = "^5.2.1"
python-dotenv = "^1.0.1"
openai = "^1.58.1"
tenacity = "^9.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
